refactor: rename file
The data-parsing module is actually just concerned with parsing strings. Renamed to better match this.
