+++ test/upm-config-io.test.ts
@@ -6,11 +6,11 @@ import {
 import { tryGetHomePath } from "../src/io/special-paths";
 import { Err, Ok } from "ts-results-es";
 import { IOError, NotFoundError, tryReadTextFromFile } from "../src/io/file-io";
-import { StringFormatError, tryParseToml } from "../src/utils/data-parsing";
+import { StringFormatError, tryParseToml } from "../src/utils/string-parsing";
 
 jest.mock("../src/io/file-io");
 jest.mock("../src/io/special-paths");
-jest.mock("../src/utils/data-parsing");
+jest.mock("../src/utils/string-parsing");
 
 describe("upm-config-io", () => {
   describe("get directory", () => {
